WIP: Extract XML-Dumping to own class
