+++ scripts/commons/serviceportal/forms/FormDumper.groovy
@@ -55,104 +55,6 @@ class FormDumper {
     this.api = api
   }
 
-  /**
-   * Dump the form as a XML string where there is a single 'serviceportal' field. This field includes a 'serviceportal-fields' sub-field, with sub-fields for each group,
-   * which then has sub-fields for each group instance in this group (starting with "instance_"), which then has
-   * sub-fields for each field.
-   *
-   * If a field is a BinaryContent, the value becomes <b>multiple</b> sub-fields "base64Data", "mimetype" and "filename"
-   *
-   * If a field is a List (e. g. a Multi-Select-Field), the value becomes <b>multiple</b> sub-fields, named
-   * <code>selectedValue</code>.
-   *
-   * Example:
-   * <pre>
-   * {@code
-   <serviceportal>
-   <serviceportal-fields>
-   <mainGroupId>
-   <instance_0>
-   <textfield>TEXTFIELD</textfield>
-   <simpleCheckbox>true</simpleCheckbox>
-   <selectOptions>VAL2</selectOptions>
-   <date>2015-07-08T00:00:00.000+02:00</date>
-   <time>1970-01-01T11:55:00.000+01:00</time>
-   <money>5.66</money>
-   <checkboxList>
-   <selectedValue>VAL1</selectedValue>
-   <selectedValue>VAL2</selectedValue>
-   </checkboxList>
-   <fileupload>
-   <base64Data>ABBREVIATED_IN_EXAMPLE</base64Data>
-   <mimetype>application/pdf</mimetype>
-   <filename>dummy.pdf</filename>
-   </fileupload>
-   </instance_0>
-   </mainGroupId>
-   </serviceportal-fields>
-   </serviceportal>}
-   * </pre>
-   *
-   * @param withMetadata A boolean that controls whether a metadata-field is added or not (default = false)
-   *
-   * @return The XML as a String
-   */
-  String dumpAsXml(boolean withMetadata = false) {
-    // Get additional data about the form itself (not just the content of the filled form)
-    FormV1 formAndMapping = api.getForm(formContent.getFormId())
-    formAndMapping.setContent(formContent)
-
-    StringWriter writer = new StringWriter()
-    MarkupBuilder xml = new MarkupBuilder(writer)
-
-    Set<String> groups = formContent.fields.keySet().collect { return it.split(":")[0] }
-
-    xml."serviceportal"() {
-      if (withMetadata) {
-        Map<String, String> metadataMap = collectMetadata()
-        // Add metadata
-        "metadata"() {
-          metadataMap.keySet().each { key ->
-            "${key}"(XmlUtil.escapeXml(metadataMap.get(key)))
-          }
-        }
-      }
-
-      "serviceportal-fields"() {
-        groups.each { group ->
-          assert group.matches("^[a-zA-Z_][\\w.-]*\$"): "Failed to create XML file. Group name '$group' is not a valid name for a XML node. Please change the group name."
-          "$group"() {
-            Set<Integer> groupInstances = formContent.fields.keySet().findAll { it.startsWith("$group:") }
-                    .collect { Integer.parseInt(it.split(":")[1]) }
-            groupInstances.each { groupInstance ->
-              // XML tags can NOT start with a number, so we need to add a prefix like "instance_" to it.
-              "instance_$groupInstance"() {
-
-                Set<String> fieldKeys = formContent.fields.keySet().findAll { it.startsWith("$group:$groupInstance:") }
-                fieldKeys.each { fullKey ->
-                  String fieldKey = fullKey.split(":")[2]
-                  assert fieldKey != null
-
-                  FormFieldV1 field = formAndMapping.getFieldInInstance(new FormFieldKeyV1(group, groupInstance, fieldKey))
-                  if (shouldRenderField(field)) {
-
-                    assert fieldKey.matches("^[a-zA-Z_][\\w.-]*\$"): "Failed to create XML file. Field name '$fieldKey' is not a valid name for a XML node. Please change the field name."
-
-                    // Add field to XML object
-                    "${fieldKey}" {
-                      mkp.yieldUnescaped(renderFieldForXmlOutput(field))
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-
-    return writer.toString()
-  }
 
   /**
    * Dump the form as a Json string. It can either be a very simplified version of the user data or more complex.
@@ -243,96 +145,5 @@ class FormDumper {
     return new JsonBuilder(fieldsAndValues).toPrettyString()
   }
 
-  private String renderFieldForXmlOutput(FormFieldV1 field) {
-    def value = field.value
 
-    if (value == null) {
-      return "" // Empty fields have empty strings
-    }
-
-    switch (value.class) {
-      case String:
-        return XmlUtil.escapeXml(value as String)
-        break
-      case Boolean:
-        return value ? "true" : "false"
-        break
-      case Date:
-        Date date = value as Date
-        SimpleDateFormat sdf = new SimpleDateFormat(iso8601Format, Locale.GERMAN)
-        return sdf.format(date)
-        break
-      case BinaryContentV1:
-        BinaryContentV1 bc = value as BinaryContentV1
-        StringWriter bcWriter = new StringWriter()
-        MarkupBuilder bcXml = new MarkupBuilder(bcWriter)
-        bcXml."base64Data"(bc.data.encodeBase64().toString())
-        bcXml."mimetype"(XmlUtil.escapeXml(bc.mimetype))
-        bcXml."filename"(XmlUtil.escapeXml(bc.uploadedFilename))
-        return bcWriter.toString()
-        break
-      case List:
-        List list = value as List
-        StringWriter listWriter = new StringWriter()
-        MarkupBuilder listXml = new MarkupBuilder(listWriter)
-        list.each {
-          listXml."selectedValue"(XmlUtil.escapeXml(it as String))
-        }
-        return listWriter.toString()
-        break
-      case BigDecimal:
-        return XmlUtil.escapeXml((value as BigDecimal).toPlainString())
-        break
-      case BinaryGeoMapContentV1:
-        // JSON format is not predictable (see https://serviceportal-community.de/153, so just output the JSON as-is)
-        return XmlUtil.escapeXml((value as BinaryGeoMapContentV1).json)
-      default:
-        api.logger.warn("FormDumper.dumpAsFlatXml does not know how to display this class '${value.class}', " + "so it defaults to toString()")
-        return value.toString()
-    }
-  }
-
-
-  /**
-   * Generates a map of metadata with entries for postfachHandleId, form id, creation date
-   * and the base64-encoded content of a binary content file named 'applicantFormAsPdf'.
-   *
-   * @return map of metadata
-   */
-  private Map<String, String> collectMetadata(){
-    Map<String, String> metadata = new HashMap<>()
-
-    // Set dev or prod api url
-    Map<String, Object> processConfig = api.getVariable("processEngineConfig", Map)
-    String portal = processConfig.get("serviceportal.environment.main-portal-host").toString().trim()
-
-    // Determine the value for startedByUser based on the portal
-    StartedByUserV1 startedByUser
-    if (portal.contains("amt24") || portal.contains("service-bw")) {
-      startedByUser = api.getVariable("startedByUser", StartedByUserV1)
-    } else {
-      StartParameterV1 startParameter = api.getVariable("startParameter", StartParameterV1)
-      startedByUser = startParameter.startedByUser
-    }
-    String postfachHandle = startedByUser.postfachHandle
-    JsonSlurper jsonSlurper = new JsonSlurper()
-    def postfachHandleMap = jsonSlurper.parseText(postfachHandle)
-    String postfachHandleId = postfachHandleMap.id
-    metadata.put("postfachHandleId", postfachHandleId)
-
-    metadata.put("formId", formContent.formId)
-
-    SimpleDateFormat formatter = new SimpleDateFormat(iso8601Format)
-    String formattedCreationDate = formatter.format(new Date())
-    metadata.put("creationDate", formattedCreationDate)
-
-    BinaryContentV1 pdf = api.getVariable("applicantFormAsPdf", BinaryContentV1)
-    if (pdf != null && pdf.data != null) {
-      metadata.put("pdfApplicantFormBase64", pdf.data.encodeBase64().toString())
-    } else {
-      metadata.put("pdfApplicantFormBase64", null)
-    }
-
-    return metadata
-  }
 }
